Add branch promotion shared workflow
In many projects that have a single owner, it can be useful to
automatically promote changes to `master` only after CI runs -- and to
do that we need a workflow that will do this logic. This takes the base
logic and abstracts it out so that this is possible, which also enables
the default branch to have proper branch protections.

This will be removed in the future when we have more users.