Use secrets directly in promotion
There appears to be an issue with passing a secret in as an input
to a reusable workflow. It's not fully clear why this seems to be an
issue, but rather than fight it, it's pretty easy to just allows the
secret to be inherited anyway since it comes from the same organization.