Set workdir on full-release
The workflow for Go releasing was missing the `workdir` input that can
be specified as an input, and this results in some divergence between
snapshot and non-snapshot. This corrects it.

Additionally, the directory for output was not being created -- and this
can lead to the action failing. This creates the directory so that
there won't be any issues once release starts.