Fix workdir use in go-release workflow
The `workdir` in the `go-release` workflow does not work as I expected;
whereas I thought this was the directory to do the work and output things
such as the `dist` directory, it actually is just the current working
directory which could way easier be done by just having a `defaults.run`
value for the work dir.

This removes that setting entirely, and instead computes the `dist`
directory by parsing the configuration file for the goreleaser tool.