Refactor go-release workflow
The go-release workflow has been failing at the attestation point due
to some easy-to-mistake configuration issues, such as having the SBOM
come out with a different path. Ideally this wouldn't _fail_ the entire
release workflow, but it currently does.

This refactors the whole process for how its done. Rather than testing
for sbom to be enabled, it instead checks for an sbom file of the same
name as the object being built.