Add docker deployment workflows
This adds a few reusable workflows for publishing docker images to
either the typical Docker registry, or to the GitHub Container Registry
(GHCR).

Additionally, a starter workflow is added that makes it easy to add a
deployment pipeline for containers.

There is some undesirable duplication here that might be able to be
fixed with composite actions in the future; this is the result of GitHub
disallowing passing secrets between jobs.