Fix docker-build-and-deploy workflow
The build-and-deploy workflow was unintentionally broken due to the
restructuring for the `version`-tag computation. The bash functionality
was expecting the `${version}` to be set, but the refactoring never sets
it before reaching the branch -- leading to an empty tag.

This not only restructures it to take this into account, it also changes
the `/` and `,` characters in the `sed` expressions to now use the `@`
character, which will never appear in expressions. This should help
ensure that the regexes are working as expected.