Publish all content from dist during go-release
The `go-release` workflow is failing due to the contents of the archive
not containing the binary at the expected place. This change updates the
process to now just upload the entire `dist` directory, since it's
already pruned before -- and it should be largely symmetric with the
content of just the tar.gz and zip files.